fix a missing constant error in unixisms, related to #7
When using a chrooted connection, and the root path of that connection
does not exist, we hit a corner case where we can't create '/'. Unfortunately
around that corner was a NameError.

We now raise a specific error `NonExistentRootError`, as you should always be
able to interact with '/', and in a recursive create situation, you should never
wind up not being able to create '/' (as it should not be missing).

Also added a spec for this condition.