complain VERY LOUDLY if close! is called on the event dispatch thread, but do not deadlock
Rather than deadlocking and causing problems, we will vomit with an error and
print a Kernel.warn message so *hopefully* the user will see one of those and
realize they've done something wrong.

The only alternative at this point would be to call exit! which would be
unconscionably horrendous, or some hack that would raise an exception (at random)
in the main thread (which would be dangerous and stupid).