MAJOR CHANGE: events now queued on the threadpool!!
This is in response to @eric's request that ZK facilitate building a concurrent
event-driven app.

There is an advantage to having a separation between the slyphon-zookeeper event dispatch
thread and the ZK callbacks: shielding users from unforseen consequences of doing
connection-related things. The threadpool will run until the user calls close!

This also means that the one (potential) downside is that the user may call close! from a
callback and deadlock :).

But hey, "welcome to threads"