Try out a new EventHandlerSubscription technique: Actors
So, stealing an idea from Celluloid, make each EventHandlerSubscription an Actor,
which will allow for greater concurrency, but will provide a more sane underlying
dispatch than the zero-locking multi-thread gangbang when you raise the threadpool
count in the current implementation > 1.

With the actor model, you have only one event dispatched to a callback at a time,
but every callback has its own queue. So you get concurrency across callbacks, and
the order in which events are delivered is preserved.

When you flip to using actors by default, the Election specs take a dump, and the specs
overall run more slowly (about 5x), but this is to be expected because the subscriptions
are *much* more heavyweight.