WeakRefs do not work the way I thought they work, use a finalizer instead
Soooo, funny thing, I thought a weakref was only garbage collected if the
object it pointed *to* was cleaned up, but it seems to be a bit less predictable
than that. Anyway, just to err on the side of caution, we register a finalizer block
to clear the fork subs (preventing a memory leak), rather than use weakrefs.