Safer state inquisitor implementation, do not touch the underlying connection
Instead of calling methods on the underlying connection (which may be in the
process of closing or in some hung state), use @last_cnx_state which is the
value of the last seen session event. This means that we're only 'connected'
after we've seen the event, however this is really not that much different
in practice.

This should prevent many deadlocks, as we're not calling down into the C layer
in an indeterminate state.